requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
sqlalchemy==2.0.29
pymysql==1.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymupdf==1.22.5
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.12.0
xgboost==1.7.5
scikit-learn==1.3.0
joblib==1.3.2
pandas==2.1.4
numpy==1.26.2
dateparser==1.1.8
python-dateutil==2.8.2
tqdm==4.66.1
regex==2024.5.15
pyyaml==6.0
bcrypt==4.0.1
playwright==1.49.0
google-genai==0.1.6
aiohttp==3.13.2